chore: move teams dir resolve conflicts
