+++ plugin-hrm-form/src/components/teamsView/SkillChip.tsx
@@ -18,7 +18,7 @@ import React from 'react';
 import { Tooltip } from '@material-ui/core';
 
 import { SkillChipStyled, SmallCheckIcon, SmallDisabledIcon } from './styles';
-import { ChipText } from '../styles';
+import { ChipText } from '../../styles';
 
 type Props = {
   skill: string;
