+++ plugin-hrm-form/src/components/teamsView/taskCardEnhancement.ts
@@ -14,8 +14,8 @@
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
 
-import { getAseloFeatureFlags, getTemplateStrings } from '../hrmConfig';
-import { getInitializedCan, PermissionActions } from '../permissions';
+import { getAseloFeatureFlags, getTemplateStrings } from '../../hrmConfig';
+import { getInitializedCan, PermissionActions } from '../../permissions';
 
 const TRUNCATED_IDENTIFIER_LENGTH = 4;
 const MAX_QUEUE_LENGTH = 13;
